Rename product -> intersect
