+++ src/04_array/intersect_hard.rb
@@ -1,0 +1,2 @@
+def intersect(hash1, hash2)
+end
