--- src/04_array/product_hard.rb
@@ -1,2 +1,0 @@
-def product(hash1, hash2)
-end
